refactor: extract AddWordForm component
